added binocoeff solution

+++ A4/src/pascalTriangleCalculator/PascalTriangleCalculator.java
@@ -2,6 +2,6 @@ package pascalTriangleCalculator;
 
 public interface PascalTriangleCalculator {
     int[] calculateRow(int rowNum);
-    int getCounter();
+    long getCounter();
     void resetCounter();
 }
